RAG_assignments/CRAG/requirements.txt
streamlit>=1.28.0
langchain>=0.1.0
langchain-community>=0.0.20
langchain-openai>=0.0.5
langchain-chroma>=0.0.6
langchain-text-splitters>=0.0.1
chromadb>=0.4.22
pypdf>=3.17.0
python-dotenv>=1.0.0
faiss-cpu>=1.7.4
requests>=2.31.0
